Refactor comment resolution modules
